auth

+++ routes/property.go
@@ -10,6 +10,7 @@ import (
 	"time"
 
 	"github.com/kataras/iris/v12"
+	"github.com/kataras/iris/v12/middleware/jwt"
 	"github.com/thanhpk/randstr"
 	"gorm.io/gorm/clause"
 )
@@ -107,6 +108,21 @@ func DeleteProperty(ctx iris.Context) {
 	params := ctx.Params()
 	id := params.Get("id")
 
+	var property models.Property
+	propertyExists := storage.DB.Find(&property, id)
+
+	if propertyExists.RowsAffected == 0 {
+		utils.CreateNotFound(ctx)
+		return
+	}
+
+	claims := jwt.Get(ctx).(*utils.AccessToken)
+
+	if property.UserID != claims.ID {
+		ctx.StatusCode(iris.StatusForbidden)
+		return
+	}
+
 	propertyDeleted := storage.DB.Delete(&models.Property{}, id)
 
 	if propertyDeleted.Error != nil {
@@ -126,6 +142,13 @@ func UpdateProperty(ctx iris.Context) {
 
 	property := GetPropertyAndAssociationsByPropertyID(id, ctx)
 	if property == nil {
+		return
+	}
+
+	claims := jwt.Get(ctx).(*utils.AccessToken)
+
+	if property.UserID != claims.ID {
+		ctx.StatusCode(iris.StatusForbidden)
 		return
 	}
